Add Profile UI

+++ DeliveryNowApp/app/src/main/java/demo1/uit/thanhduyen/imageslider/Product.java
@@ -46,7 +46,7 @@ public class Product {
         return distance;
     }
 
-    public void setDistance(int distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
 }
